pyproject.toml
[project]
name = "es-orchestrator-mcp"
version = "0.1.0"
description = "MCP server orchestrating Expert Sleepers FH-2, ES-9, and Disting NT Eurorack modules"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-rtmidi>=1.5.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
]

[project.scripts]
es-orchestrator-mcp = "server:main"
